requirements.txt
absl-py==2.3.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.0.1
flatbuffers==25.2.10
gast==0.4.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.73.0
grpcio-status==1.48.2
gunicorn==20.1.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.8
markdown2==2.4.12
MarkupSafe==3.0.2
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
Pillow==9.3.0
proto-plus==1.26.1
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==0.21.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==2.0.3
wrapt==1.17.2